+++ gulpfile.js
@@ -482,6 +482,7 @@ gulp.task('js-merge', function () {
     './src/832xlsxml.js',
     './src/84from.js',
     './src/843xml.js',
+    './src/844gexf.js',
     './src/85help.js',
     './src/86print.js',
     './src/87source.js',